# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "3CM API Blog",
        "description": "API documentation for the 3CM Blog project",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/comments": {
            "get": {
                "tags": [
                    "Admin Comments"
                ],
                "summary": "List all comments (Admin)",
                "description": "Display a listing of comments (Admin).",
                "operationId": "429f8d249f5808f531aaac718f410bb7",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Comment"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/articles/{id}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get comments for an article",
                "description": "Get comments for an article.",
                "operationId": "809bad0d46bd96101b6c848e56f43aa8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Comment"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Submit a comment for an article",
                "description": "Store a newly created comment in storage.",
                "operationId": "dd250b72b3cbbf26b50ef20800d84c97",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "author_name",
                                    "content"
                                ],
                                "properties": {
                                    "author_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "author_avatar": {
                                        "type": "string",
                                        "example": "https://example.com/avatar.png"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Great article!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/comments/{id}/approve": {
            "put": {
                "tags": [
                    "Admin Comments"
                ],
                "summary": "Approve a comment (Admin)",
                "description": "Approve a comment (Admin).",
                "operationId": "a2ec14e53859e6f5a01463edfc6c3fdb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment approved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/comments/{id}": {
            "delete": {
                "tags": [
                    "Admin Comments"
                ],
                "summary": "Delete a comment (Admin)",
                "description": "Remove the specified comment from storage (Admin).",
                "operationId": "0b9972707c6ea29998662301eb0fe9c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/events/{id}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get comments for an event",
                "description": "Get comments for an event.",
                "operationId": "148cd10cdd59a3c2b17736cf29320e64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Comment"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Submit a comment for an event",
                "description": "Store a newly created comment for an event.",
                "operationId": "ed3b80bf8d942a909f6c85b8270f7446",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "author_name",
                                    "content"
                                ],
                                "properties": {
                                    "author_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "author_avatar": {
                                        "type": "string",
                                        "example": "https://example.com/avatar.png"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Great event!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Comment": {
                "required": [
                    "commentable_type",
                    "commentable_id",
                    "author_name",
                    "content"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "commentable_type": {
                        "type": "string"
                    },
                    "commentable_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "author_name": {
                        "type": "string"
                    },
                    "author_avatar": {
                        "type": "string",
                        "nullable": true
                    },
                    "content": {
                        "type": "string"
                    },
                    "is_approved": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}